requirements.txt
beautifulsoup4
requests
python-dotenv
openai
jupyterlab